Refactor query utils
