+++ src/routes/details/ociDetails/ociDetails.tsx
@@ -21,6 +21,8 @@ import { getIdKeyForGroupBy } from 'routes/utils/computedReport/getComputedOciRe
 import type { ComputedReportItem } from 'routes/utils/computedReport/getComputedReportItems';
 import { getUnsortedComputedReportItems } from 'routes/utils/computedReport/getComputedReportItems';
 import { getGroupByTagKey } from 'routes/utils/groupBy';
+import { filterProviders, hasCurrentMonthData } from 'routes/utils/providers';
+import { getRouteForQuery } from 'routes/utils/query';
 import {
   handleOnCurrencySelected,
   handleOnFilterAdded,
@@ -28,9 +30,7 @@ import {
   handleOnPerPageSelect,
   handleOnSetPage,
   handleOnSort,
-} from 'routes/utils/navHandles';
-import { filterProviders, hasCurrentMonthData } from 'routes/utils/providers';
-import { getRouteForQuery } from 'routes/utils/query';
+} from 'routes/utils/queryNavigate';
 import { createMapStateToProps, FetchStatus } from 'store/common';
 import { providersQuery, providersSelectors } from 'store/providers';
 import { reportActions, reportSelectors } from 'store/reports';
