+++ src/routes/details/ocpBreakdown/optimizationsBreakdown.tsx
@@ -15,9 +15,9 @@ import { NoOptimizations } from 'routes/components/page/noOptimizations';
 import { NotAvailable } from 'routes/components/page/notAvailable';
 import { styles } from 'routes/optimizations/optimizations.styles';
 import { getGroupById, getGroupByValue } from 'routes/utils/groupBy';
-import { handleOnFilterAdded, handleOnFilterRemoved, handleOnSort } from 'routes/utils/navHandles';
 import { getOrderById, getOrderByValue } from 'routes/utils/orderBy';
 import { getRouteForQuery } from 'routes/utils/query';
+import { handleOnFilterAdded, handleOnFilterRemoved, handleOnSort } from 'routes/utils/queryNavigate';
 import { createMapStateToProps, FetchStatus } from 'store/common';
 import { rosActions, rosSelectors } from 'store/ros';
 import type { RouterComponentProps } from 'utils/router';
