+++ src/routes/settings/costCategory/costCategory.tsx
@@ -14,7 +14,7 @@ import { Loading } from 'routes/components/page/loading';
 import { NotAvailable } from 'routes/components/page/notAvailable';
 import type { ComputedReportItem } from 'routes/utils/computedReport/getComputedReportItems';
 import { getUnsortedComputedReportItems } from 'routes/utils/computedReport/getComputedReportItems';
-import * as handles from 'routes/utils/handles';
+import * as queryUtils from 'routes/utils/query';
 import type { RootState } from 'store';
 import { FetchStatus } from 'store/common';
 import { reportActions, reportSelectors } from 'store/reports';
@@ -159,22 +159,22 @@ const CostCategory: React.FC<CostCategoryProps> = () => {
   const handleOnEnableCategories = () => {};
 
   const handleOnFilterAdded = filter => {
-    const newQuery = handles.handleOnFilterAdded(query, filter);
+    const newQuery = queryUtils.handleOnFilterAdded(query, filter);
     setQuery(newQuery);
   };
 
   const handleOnFilterRemoved = filter => {
-    const newQuery = handles.handleOnFilterRemoved(query, filter);
+    const newQuery = queryUtils.handleOnFilterRemoved(query, filter);
     setQuery(newQuery);
   };
 
   const handleOnPerPageSelect = perPage => {
-    const newQuery = handles.handleOnPerPageSelect(query, perPage);
+    const newQuery = queryUtils.handleOnPerPageSelect(query, perPage);
     setQuery(newQuery);
   };
 
   const handleOnSetPage = pageNumber => {
-    const newQuery = handles.handleOnSetPage(query, report, pageNumber);
+    const newQuery = queryUtils.handleOnSetPage(query, report, pageNumber);
     setQuery(newQuery);
   };
 
@@ -194,7 +194,7 @@ const CostCategory: React.FC<CostCategoryProps> = () => {
   };
 
   const handleOnSort = (sortType, isSortAscending) => {
-    const newQuery = handles.handleOnSort(query, sortType, isSortAscending);
+    const newQuery = queryUtils.handleOnSort(query, sortType, isSortAscending);
     setQuery(newQuery);
   };
